double naar int

+++ src/main/java/com/proj2/model/Travel.java
@@ -5,9 +5,9 @@ import com.proj2.model.abstraction.AbstractVehicle;
 public class Travel {
     private int points;
 
-    public Travel(AbstractVehicle vehicle, double aantalKM) {
+    public Travel(AbstractVehicle vehicle, int aantalKM) {
         // formule voor het berekenen van punten
-        points = aantalKM * vehicle.getModifier();
+        points = (int) (aantalKM * vehicle.getModifier());
     }
 
     public int getPoints() {
